pyproject.toml
[tool.poetry]
name = "flumes-fuse"
version = "0.1.0"
description = ""
authors = ["Jorge Zapata <jorgeluis.zapata@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fuse-python = "^1.0.4"
SQLAlchemy = "^1.4.26"
flumes = "^0.1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.10b0"
isort = "^5.10.0"
pre-commit = "^2.15.0"

[tool.poetry.scripts]
flumes-fuse = "flumes_fuse.fs:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
